+++ src/web/about.rs
@@ -1,9 +1,26 @@
-use yew::prelude::*;
+use super::logs::Logs;
+use crate::*;
+use yew::{prelude::*, virtual_dom::VNode};
+use yew_export_button::{export_button, ButtonOpts};
 
 const VERSION: &'static str = env!("CARGO_PKG_VERSION");
 const REPO_URL: &str = "https://github.com/Terkwood/equanimity";
 
-pub fn section(callback: Callback<MouseEvent>) -> Html {
+const EXPORT_BUTTON_CSS_ID: &str = "export-button";
+const EXPORT_LINK_CSS_CLASS: &str = "fancy-button thick";
+const EXPORT_FILE_PREFIX: &str = "equanimity";
+
+pub fn section(callback: Callback<MouseEvent>, ctx: &yew::Context<Logs>) -> Html {
+    let export_button: VNode = export_button(
+        &ctx.props().storage_state,
+        ButtonOpts {
+            a_class: EXPORT_LINK_CSS_CLASS.to_string(),
+            button_id: EXPORT_BUTTON_CSS_ID.to_string(),
+            file_prefix: EXPORT_FILE_PREFIX.to_string(),
+            utc_millis: utc_now(),
+        },
+    );
+
     html! {
         <div id="about">
             <h1>{ "About" }</h1>
@@ -13,6 +30,11 @@ pub fn section(callback: Callback<MouseEvent>) -> Html {
             <h2>{ "Source Code" }</h2>
             <p>{ "The source code is available under MIT license." }</p>
             <p><a href={REPO_URL}>{ REPO_URL }</a></p>
+
+            <div class="center">
+                {  export_button }
+            </div>
+
             <button
                 class="fancy-button thick"
                 role="button"
